update other tests
