+++ test/algorithms/legacy/test_min_eigen_optimizer.py
@@ -43,6 +43,7 @@ from qiskit_optimization.converters import (
 )
 from qiskit_optimization.problems import QuadraticProgram
 from qiskit_optimization.runtime import VQEProgram, QAOAProgram
+from qiskit_optimization.deprecation import clear_deprecated_objects
 
 
 @ddt
@@ -381,6 +382,15 @@ class TestMinEigenOptimizer(QiskitOptimizationTestCase):
         opt = MinimumEigenOptimizer(solver)
         result = opt.solve(self.op_ordering)
         self.assertIsInstance(result, MinimumEigenOptimizationResult)
+
+    def test_deprecation(self):
+        """Test deprecation warning"""
+        clear_deprecated_objects()
+        optimizer = SPSA(maxiter=100)
+        ry_ansatz = TwoLocal(5, "ry", "cz", reps=3, entanglement="full")
+        vqe_mes = VQE(ry_ansatz, optimizer=optimizer, quantum_instance=QasmSimulatorPy())
+        with self.assertWarns(DeprecationWarning):
+            _ = MinimumEigenOptimizer(vqe_mes)
 
 
 if __name__ == "__main__":
