+++ test/algorithms/test_min_eigen_optimizer.py
@@ -19,7 +19,6 @@ from test.runtime.fake_vqeruntime import FakeQAOARuntimeProvider, FakeVQERuntime
 import numpy as np
 from ddt import data, ddt, unpack
 from qiskit.algorithms.minimum_eigensolvers import QAOA, NumPyMinimumEigensolver, SamplingVQE
-from qiskit.algorithms.minimum_eigen_solvers import VQE
 from qiskit.algorithms.optimizers import COBYLA, SPSA
 from qiskit.circuit.library import TwoLocal
 from qiskit.primitives import Sampler
@@ -363,14 +362,6 @@ class TestMinEigenOptimizer(QiskitOptimizationTestCase):
         opt = MinimumEigenOptimizer(solver)
         result = opt.solve(self.op_ordering)
         self.assertIsInstance(result, MinimumEigenOptimizationResult)
-
-    def test_deprecation(self):
-        """Test deprecation warning"""
-        optimizer = SPSA(maxiter=100)
-        ry_ansatz = TwoLocal(5, "ry", "cz", reps=3, entanglement="full")
-        vqe_mes = VQE(ry_ansatz, optimizer=optimizer, quantum_instance=QasmSimulatorPy())
-        with self.assertWarns(DeprecationWarning):
-            _ = MinimumEigenOptimizer(vqe_mes)
 
 
 if __name__ == "__main__":
